created front route /judge/innovation/[id] that is hydrated with the database
