+++ app/layout.tsx
@@ -7,8 +7,8 @@ const defaultUrl = process.env.VERCEL_URL
 
 export const metadata = {
   metadataBase: new URL(defaultUrl),
-  title: "Next.js and Supabase Starter Kit",
-  description: "The fastest way to build apps with Next.js and Supabase",
+  title: "FLL Ranker",
+  description: "First Lego League team ranking toolkit",
 };
 
 export default function RootLayout({
