+++ components/forms/evaluationGrid.tsx
@@ -7,9 +7,11 @@ import { createClient } from "@/utils/supabase/client";
 export default function EvaluationGrid({
   evalAxes,
   data,
+  teamId
 }: {
   evalAxes: EvalAxis[];
   data: EvaluationDB | null;
+  teamId :number;
 }) {
   const initialState: EvaluationState = {
     identify: [
@@ -95,7 +97,7 @@ export default function EvaluationGrid({
     const { error } = await supabase
       .from("innovation_scoresheet")
       .update(updatedEvaluation)
-      .eq("team_id", 1);
+      .eq("team_id", teamId);
 
     console.log(error?.code);
   };
@@ -130,7 +132,7 @@ export default function EvaluationGrid({
       <button
         type="submit"
         onClick={handleSubmit}
-        className="m-8 focus:outline-none text-white bg-green-700 hover:bg-green-800 focus:ring-4 focus:ring-green-300 font-medium rounded-lg text-sm px-5 py-2.5 me-2 mb-2 dark:bg-green-600 dark:hover:bg-green-700 dark:focus:ring-green-800"
+        className="m-8 focus:outline-none text-white bg-green-700 hover:bg-green-800 focus:ring-4 focus:ring-green-300 font-medium rounded-lg text-sm px-5 py-2.5 me-2 mb-2 "
       >
         Submit
       </button>
